+++ src/cpp_bindings/dense.cpp
@@ -152,7 +152,8 @@ void init_dense(py::module_ &module_matrix)
                                                                 struct-style
                                                                 format
                                                                 descriptor */
-                    dim,          /* Number of dimensions */
+                    dim, /* Number of dimensions */
+                    // TODO: potential mistake:
                     {rows, cols}, /* Buffer dimensions */
                     {sizeof(ValueType), sizeof(ValueType) * rows}
                     /* Strides (in bytes) for each index */
@@ -179,10 +180,6 @@ void init_dense(py::module_ &module_matrix)
                  o->fill(s);
                  m.inv_scale(o);
              })
-        // TODO: those functions were actually newer tested.
-        // With the following changes they should be working now,
-        // yet the tests for them should be definitely added in
-        // the following PR
         .def(
             "add_scaled",
             [](gko::matrix::Dense<ValueType> &self,