+++ src/cpp_bindings/matrix.cpp
@@ -23,26 +23,26 @@ namespace py = pybind11;
                 vals, cols, rows));                                           \
         }))                                                                   \
         .def(py::init([](std::shared_ptr<gko::Executor> exec, py::tuple dim,  \
-                         py::buffer data, py::buffer rows, py::buffer cols) { \
+                         py::buffer data, py::buffer cols, py::buffer rows) { \
             /* Request a buffer descriptor from Python */                     \
             py::buffer_info data_info = data.request();                       \
-            py::buffer_info rows_info = rows.request();                       \
             py::buffer_info cols_info = cols.request();                       \
+            py::buffer_info rows_info = rows.request();                       \
                                                                               \
             if (data_info.format !=                                           \
                 py::format_descriptor<ValueType>::format())                   \
                 throw std::runtime_error(                                     \
                     "Provided values have an incompatible dtype");            \
                                                                               \
-            if (rows_info.format !=                                           \
-                py::format_descriptor<IndexType>::format())                   \
-                throw std::runtime_error(                                     \
-                    "Provided rows have an incompatible dtype");              \
-                                                                              \
             if (cols_info.format !=                                           \
                 py::format_descriptor<IndexType>::format())                   \
                 throw std::runtime_error(                                     \
                     "Provided cols have an incompatible dtype");              \
+                                                                              \
+            if (rows_info.format !=                                           \
+                py::format_descriptor<IndexType>::format())                   \
+                throw std::runtime_error(                                     \
+                    "Provided rows have an incompatible dtype");              \
                                                                               \
             auto ref = gko::ReferenceExecutor::create();                      \
                                                                               \
@@ -51,22 +51,17 @@ namespace py = pybind11;
                                                                               \
             auto data_view = gko::array<ValueType>::view(                     \
                 ref, nnz, (ValueType *)data_info.ptr);                        \
-            auto rows_view = gko::array<IndexType>::view(                     \
-                ref, nnz, (IndexType *)rows_info.ptr);                        \
             auto cols_view = gko::array<IndexType>::view(                     \
                 ref, nnz, (IndexType *)cols_info.ptr);                        \
+                                                                              \
+            auto rows_view = gko::array<IndexType>::view(                     \
+                ref, rows_info.shape[0], (IndexType *)rows_info.ptr);         \
                                                                               \
             return gko::share(gko::matrix::Name<ValueType>::create(           \
                 exec,                                                         \
                 gko::dim<2>{dim[0].cast<size_t>(), dim[1].cast<size_t>()},    \
                 data_view, cols_view, rows_view));                            \
         }))                                                                   \
-        .def(                                                                 \
-            "apply",                                                          \
-            [](const gko::matrix::Name<ValueType, IndexType> &m,              \
-               std::shared_ptr<const gko::LinOp> b,                           \
-               std::shared_ptr<gko::LinOp> x) { m.apply(b, x); },             \
-            "")                                                               \
         .def("__repr__",                                                      \
              [](const gko::matrix::Name<ValueType> &o) {                      \
                  auto str = std::string("pygko.matrix.Name object");          \
@@ -75,8 +70,6 @@ namespace py = pybind11;
         .def("get_num_stored_elements",                                       \
              &gko::matrix::Name<ValueType>::get_num_stored_elements,          \
              "Get the number of non zero elements")                           \
-        .def("get_executor", &gko::matrix::Name<ValueType>::get_executor,     \
-             "Get the executor")                                              \
         .def("get_size", &gko::matrix::Name<ValueType>::get_size,             \
              "Get the size of the matrix")