+++ tests/cpp_bindings/test_cuda_executor.py
@@ -3,27 +3,40 @@
 # SPDX-FileCopyrightText: 2024 pyGinkgo authors
 
 import pytest
-import sys
 import numpy as np
 
-sys.path.append("../../")
-import pyGinkgoBindings as pGB
+import pyGinkgo.pyGinkgoBindings as pGB
+
+from test_utils import verify_dense_vec
 
 
-@pytest.mark.gpu
 def test_array_cuda():
     if pGB.CudaExecutor.get_num_devices() < 1:
         pytest.skip("CUDA is not available")
 
     executor = pGB.CudaExecutor()
     np_array = np.array([1.0, 2.0, 3.0, 4.0, 5.0])
-    arr = pGB.base.array(executor, np_array)
-    arr_copy = pGB.base.array(executor, arr)
+    arr = pGB.base.array_double(executor, np_array)
+    arr_copy = pGB.base.array_double(executor, arr)
     assert arr.get_size() == arr_copy.get_size()
     assert pGB.base.reduce_add(arr, 0.0) == 15.0
 
 
-@pytest.mark.gpu
+def test_dense_copy_to_host():
+    if pGB.CudaExecutor.get_num_devices() < 1:
+        pytest.skip("CUDA is not available")
+
+    master = pGB.ReferenceExecutor()
+    executor = pGB.CudaExecutor(master=master)
+    np_array = np.array([1.0, 2.0, 3.0, 4.0, 5.0])
+    dense = pGB.matrix.dense(executor, np_array)
+    dense_on_master = dense.copy_to_host()
+    assert dense.get_executor() == executor
+    assert dense_on_master.get_executor() == master
+    assert id(dense) != id(dense_on_master)
+    verify_dense_vec(dense_on_master, np_array)
+
+
 def test_device_id():
     if pGB.CudaExecutor.get_num_devices() < 1:
         pytest.skip("CUDA is not available")
@@ -34,7 +47,6 @@ def test_device_id():
         executor.device_id = 1
 
 
-@pytest.mark.gpu
 def test_master():
     if pGB.CudaExecutor.get_num_devices() < 1:
         pytest.skip("CUDA is not available")